Update build_loop.yml
Adding CGM Clients to 3.2.3 using Browser
This method was updated 05 January 2024 to accomodate changes in one of the client repositories. You will need to update your build_loop.yml file for this to continue working. The change to the client repository happened 12 November 2023.

The prepared fork at loopnlearn used to download Loop with Patches will not work with Xcode 15.

If you are using the Browser Build and want to add a CGM Client, then follow the directions below and paste those lines into build_loop.yml

Warning – This will not work with the next release of Loop (3.4.0) and does not work with Loop dev.